requirements.txt
uvicorn
fastapi
opencv-python
python-dotenv